+++ src/tools/ta_regtest/ta_variant_frame.h
@@ -3949,6 +3949,45 @@ static TA_RetCode TA_COSH_VFrameS( int startIdx, int endIdx,
 static const TA_VInputKind TA_VIn_COSH[] = { TA_VIN_REAL };
 static const int TA_VOutIsInt_COSH[] = { 0 };
 
+static TA_RetCode TA_CTI_VFrameD( int startIdx, int endIdx,
+                  const double *const in[], const double optIn[],
+                  int *outBegIdx, int *outNBElement,
+                  double *const outReal[], int *const outInteger[] )
+{
+   (void)outInteger;
+   return TA_CTI(
+               startIdx,
+               endIdx,
+               in[0] /* inReal */,
+               (int)optIn[0] /* optInTimePeriod */,
+               outBegIdx,
+               outNBElement,
+               outReal[0] /* outReal */
+               );
+}
+static TA_RetCode TA_CTI_VFrameS( int startIdx, int endIdx,
+                  const float *const in[], const double optIn[],
+                  int *outBegIdx, int *outNBElement,
+                  double *const outReal[], int *const outInteger[] )
+{
+   (void)outInteger;
+   return TA_S_CTI(
+               startIdx,
+               endIdx,
+               in[0] /* inReal */,
+               (int)optIn[0] /* optInTimePeriod */,
+               outBegIdx,
+               outNBElement,
+               outReal[0] /* outReal */
+               );
+}
+
+static const TA_VInputKind TA_VIn_CTI[] = { TA_VIN_REAL };
+static const int TA_VOutIsInt_CTI[] = { 0 };
+static const TA_VOptSpec TA_VOpt_CTI[] = {
+   { "optInTimePeriod", TA_VOPT_INT, 2.0, 100000.0, 20.0 },
+};
+
 static TA_RetCode TA_CUMSUM_VFrameD( int startIdx, int endIdx,
                   const double *const in[], const double optIn[],
                   int *outBegIdx, int *outNBElement,
@@ -8680,6 +8719,8 @@ static const TA_VariantEntry TA_VariantTable[] = {
      1, TA_VIn_COS, 0, NULL, 1, TA_VOutIsInt_COS, 0 },
    { "COSH", TA_COSH_VFrameD, TA_COSH_VFrameS,
      1, TA_VIn_COSH, 0, NULL, 1, TA_VOutIsInt_COSH, 0 },
+   { "CTI", TA_CTI_VFrameD, TA_CTI_VFrameS,
+     1, TA_VIn_CTI, 1, TA_VOpt_CTI, 1, TA_VOutIsInt_CTI, 0 },
    { "CUMSUM", TA_CUMSUM_VFrameD, TA_CUMSUM_VFrameS,
      1, TA_VIn_CUMSUM, 0, NULL, 1, TA_VOutIsInt_CUMSUM, 0 },
    { "CVI", TA_CVI_VFrameD, TA_CVI_VFrameS,
@@ -8902,6 +8943,6 @@ static const TA_VariantEntry TA_VariantTable[] = {
      1, TA_VIn_ZLEMA, 1, TA_VOpt_ZLEMA, 1, TA_VOutIsInt_ZLEMA, 0 },
 };
 
-#define TA_VARIANT_TABLE_SIZE 201
+#define TA_VARIANT_TABLE_SIZE 202
 
 #endif /* TA_VARIANT_FRAME_H */